backend/pyproject.toml
[project]
name = "slopit"
version = "0.1.0"
description = "Reddit-inspired social network with three-level random feed algorithm"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # Core
    "django>=5.1,<5.2",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.4",
    "django-filter>=24.3",
    "django-environ>=0.11",
    "python-dotenv>=1.0",
    "psycopg[binary]>=3.2",
    "dj-database-url>=2.2",

    # Auth
    "django-allauth[socialaccount]>=65.0",
    "fido2>=1.1",              # WebAuthn / passkey support for allauth.mfa
    "PyJWT>=2.9",
    "cryptography>=43.0",
    "python-telegram-bot>=21.0",  # Telegram bot server (runbot management command)

    # Admin & utilities
    "django-unfold>=0.40",
    "django-extensions>=3.2",
    "pillow>=10.4",
    "markdown>=3.7",
    "bleach>=6.1",

    # Background jobs (used in Stage 3 for feed regeneration)
    "django-rq>=3.0",
    "redis>=5.1",

    # API schema
    "drf-spectacular>=0.27",

    # Observability
    "sentry-sdk>=2.14",  # GlitchTip is Sentry-compatible

    # Production server
    "gunicorn>=23.0",
    "whitenoise>=6.7",

    # Cloud media storage (Supabase Storage via S3-compatible API)
    "django-storages[s3]>=1.14",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6",
    "mypy>=1.11",
    "django-stubs[compatible-mypy]>=5.1",
    "djangorestframework-stubs[compatible-mypy]>=3.15",
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=5.0",
    "pytest-xdist>=3.6",
    "playwright>=1.54",
    "factory-boy>=3.3",
    "faker>=28.0",
    "django-debug-toolbar>=4.4",
    "ipython>=8.27",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slopit"]

# ─── Ruff ────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src/slopit"]
cache-dir = ".ruff_cache"
extend-exclude = ["**/migrations/*"]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "N",        # pep8-naming
    "UP",       # pyupgrade
    "B",        # bugbear
    "C4",       # comprehensions
    "DJ",       # django specific
    "S",        # bandit (security)
    "SIM",      # simplify
    "RET",      # return
    "TC",       # type-checking
    "PTH",      # pathlib
    "RUF",      # ruff-specific
]
ignore = [
    "S101",  # assert allowed in tests
    "DJ001", # nullable CharField — sometimes needed
    "RUF012", # ClassVar annotation — too noisy for Django Model/Admin meta classes
]

[tool.ruff.per-file-ignores]
"tests/**/*" = ["S101", "S106", "PLR2004"]
"**/settings/*" = ["S105", "F403", "F405"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ─── mypy ────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
warn_redundant_casts = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
exclude = ["migrations/"]

[tool.django-stubs]
django_settings_module = "config.settings.dev"

[[tool.mypy.overrides]]
module = ["allauth.*", "django_rq.*", "unfold.*"]
ignore_missing_imports = true

# ─── pytest ──────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py", "*_test.py"]
pythonpath = ["src/slopit"]
cache_dir = ".pytest_cache"
addopts = [
    "--strict-markers",
    "--reuse-db",
    "-ra",
    "--cov=src/slopit",
    "--cov-report=term-missing",
    "--cov-fail-under=60",
]
markers = [
    "slow: marks tests as slow",
    "integration: integration tests requiring DB",
]

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/manage.py",
    "*/settings/*",
    "*/wsgi.py",
    "*/asgi.py",
]
